Add WakeLockControl components
