+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export { default as WarningModal } from './WarningModal.vue'
 export { default as WakeLockToggle } from './WakeLockToggle.vue'
+export { default as WakeLockControl } from './controls/WakeLockControl.vue'
