+++ src/views/HomeView.vue
@@ -27,7 +27,7 @@
         class="mb-8"
       >
         <h2 id="wake-lock-section" class="sr-only">ウェイクロック制御</h2>
-        <WakeLockToggle @request-wake-lock="handleWakeLockRequest" />
+        <WakeLockControl />
       </section>
       
       <!-- Info Section -->
@@ -65,63 +65,9 @@
       </div>
     </div>
     
-    <!-- Warning Modal -->
-    <WarningModal 
-      :is-visible="showWarningModal"
-      :is-processing="isModalProcessing"
-      @confirm="handleModalConfirm"
-      @close="handleModalClose"
-    />
   </div>
 </template>
 
 <script setup lang="ts">
-import { ref } from 'vue';
-import { WakeLockToggle, WarningModal } from '../components';
-
-// Modal state management
-const showWarningModal = ref(false);
-const hasShownWarning = ref(false);
-const isModalProcessing = ref(false);
-
-// Pending wake lock request callback
-let pendingWakeLockRequest: (() => Promise<void>) | null = null;
-
-const handleWakeLockRequest = async (requestCallback: () => Promise<void>) => {
-  if (!hasShownWarning.value) {
-    pendingWakeLockRequest = requestCallback;
-    showWarningModal.value = true;
-    return;
-  }
-  
-  try {
-    await requestCallback();
-  } catch (error) {
-    console.error('Wake lock request failed:', error);
-  }
-};
-
-const handleModalConfirm = async () => {
-  if (isModalProcessing.value) return;
-
-  isModalProcessing.value = true;
-  hasShownWarning.value = true;
-  
-  try {
-    if (pendingWakeLockRequest) {
-      await pendingWakeLockRequest();
-    }
-  } catch (error) {
-    console.error('Wake lock request failed after modal confirmation:', error);
-  } finally {
-    showWarningModal.value = false;
-    isModalProcessing.value = false;
-    pendingWakeLockRequest = null;
-  }
-};
-
-const handleModalClose = () => {
-  showWarningModal.value = false;
-  pendingWakeLockRequest = null;
-};
+import { WakeLockControl } from '../components';
 </script>
